refactor: un-share blocks
We had a couple blocks and attributes in the shared file that weren't
actually shared.
