+++ internal/provider/component_helm_chart_resource.go
@@ -80,7 +80,21 @@ func (r *HelmChartComponentResource) Schema(ctx context.Context, req resource.Sc
 			"connected_repo": connectedRepoAttribute(),
 		},
 		Blocks: map[string]schema.Block{
-			"value": helmValueSharedBlock(),
+			"value": schema.ListNestedBlock{
+				Description: "Environment variables to export into the env when running the image.",
+				NestedObject: schema.NestedBlockObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							Description: "The variable name to export to the env (e.g. API_TOKEN or PORT.)",
+							Required:    true,
+						},
+						"value": schema.StringAttribute{
+							Description: "The variable value to export to the env. Can be any valid env var value, or interpolated from Nuon.",
+							Required:    true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
